@@ -11,7 +11,7 @@
     },
     "devDependencies": {
         "axios": "^0.18",
-        "bootstrap": "^4.0.0",
+        "bootstrap": "^4.3.1",
         "cross-env": "^5.1",
         "jquery": "^3.2",
         "laravel-mix": "^4.0.7",
@@ -21,10 +21,11 @@
         "sass": "^1.15.2",
         "sass-loader": "^7.1.0",
         "swagger-ui": "^3.20.5",
-        "vue": "^2.5.17",
+        "vue": "^2.6.6",
         "vue-template-compiler": "^2.5.22"
     },
     "dependencies": {
+        "bootstrap-vue": "^2.0.0-rc.12",
         "vue-router": "^3.0.2"
     }
 }
