@@ -20,22 +20,19 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-/**
- * First we will load all of this project's JavaScript dependencies which
- * includes Vue and other libraries. It is a great starting point when
- * building robust, powerful web applications using Vue and Laravel.
- */
-
 require('./bootstrap');
 
-window.Vue = require('vue');
+import Vue from "vue";
 import VueRouter from "vue-router";
+import BootstrapVue from "bootstrap-vue";
 
-Vue.use(VueRouter)
+Vue.use(VueRouter);
+Vue.use(BootstrapVue);
 
 import App from './views/App'
 import Hello from './views/Hello'
 import Home from './views/Home'
+import Events from './views/Events'
 
 /**
  * The following block of code may be used to automatically register your
@@ -57,6 +54,12 @@ const router = new VueRouter({
             path: '/hello',
             name: 'hello',
             component: Hello,
+        },
+
+        {
+            path: '/events',
+            name: 'events',
+            component: Events,
         },
     ],
 });
