@@ -20,18 +20,52 @@
   -->
 
 <template>
+
     <div>
-        <h1>Vue Router Demo App</h1>
 
-        <p>
-            <router-link :to="{ name: 'home' }">Home</router-link> |
-            <router-link :to="{ name: 'hello' }">Hello World</router-link>
-        </p>
+        <b-navbar toggleable="lg">
+            <b-navbar-brand href="#">CatLab Drinks</b-navbar-brand>
 
-        <div class="container">
-            <router-view></router-view>
-        </div>
+            <b-navbar-toggle target="nav_collapse" />
+
+            <b-collapse is-nav id="nav_collapse">
+                <b-navbar-nav>
+                    <b-nav-item :to="{ name: 'home' }">Home</b-nav-item>
+                    <b-nav-item :to="{ name: 'hello' }">Hello world</b-nav-item>
+                </b-navbar-nav>
+
+                <!-- Right aligned nav items -->
+                <b-navbar-nav class="ml-auto">
+
+                    <b-nav-item :to="{ name: 'events' }">Events</b-nav-item>
+
+                    <!--
+                    <b-nav-form>
+                        <b-form-input size="sm" class="mr-sm-2" type="text" placeholder="Search" />
+                        <b-button size="sm" class="my-2 my-sm-0" type="submit">Search</b-button>
+                    </b-nav-form>
+
+                    <b-nav-item-dropdown text="Lang" right>
+                        <b-dropdown-item href="#">EN</b-dropdown-item>
+                        <b-dropdown-item href="#">ES</b-dropdown-item>
+                        <b-dropdown-item href="#">RU</b-dropdown-item>
+                        <b-dropdown-item href="#">FA</b-dropdown-item>
+                    </b-nav-item-dropdown>
+
+                    <b-nav-item-dropdown right>
+                        <template slot="button-content"><em>User</em></template>
+                        <b-dropdown-item href="#">Profile</b-dropdown-item>
+                        <b-dropdown-item href="#">Signout</b-dropdown-item>
+                    </b-nav-item-dropdown>
+                    -->
+                </b-navbar-nav>
+            </b-collapse>
+        </b-navbar>
+
+        <router-view></router-view>
+
     </div>
+
 </template>
 <script>
     export default {}
