@@ -2,11 +2,18 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Vue SPA Demo</title>
+    <title>CatLab Drinks</title>
 
     <base href="/client/" />
+
+    <!-- CSRF Token -->
+    <meta name="csrf-token" content="{{ csrf_token() }}">
+
+    <!-- Styles -->
+    <link href="{{ asset('client-res/css/app.css') }}" rel="stylesheet">
 </head>
 <body>
 <div id="app">
